// src/redux/createEntityPageContent/createEntityPageContent.types.ts
export interface HeaderPageContent {
  title: string
  shortDescription: string
  imageDescription: string
  sdgs: string[]
  brand: string
  location: string
  headerFileSrc: string
  headerFileUploading: boolean
  logoFileSrc: string
  logoFileUploading: boolean
}

export interface BodyPageContent {
  id: string
  title: string
  content: string
  fileSrc: string
  uploading: boolean
}

export interface ImagePageContent {
  id: string
  title: string
  content: string
  imageDescription: string
  fileSrc: string
  uploading: boolean
}

export interface VideoPageContent {
  id: string
  title: string
  content: string
  fileSrc: string
  uploading: boolean
}

export interface ProfilePageContent {
  id: string
  name: string
  position: string
  linkedInUrl: string
  twitterUrl: string
  fileSrc: string
  uploading: boolean
}

export interface SocialPageContent {
  linkedInUrl: string
  facebookUrl: string
  twitterUrl: string
  discourseUrl: string
  instagramUrl: string
  telegramUrl: string
  githubUrl: string
  otherUrl: string
}

export interface EmbeddedPageContent {
  id: string
  title: string
  urls: string[]
}

export interface CreateEntityPageContentState {
  header: HeaderPageContent
  body: {
    [id: string]: BodyPageContent
  }
  images: {
    [id: string]: ImagePageContent
  }
  profiles: {
    [id: string]: ProfilePageContent
  }
  social: SocialPageContent
  embedded: {
    [id: string]: EmbeddedPageContent
  }
  validation: {
    [identifier: string]: Validation
  }
}